Refactor transport namespace
- Deprecate non-DSN configurations
- Break each transport into its own module space
- Generically handle querystrings-as-options
