+++ raven/transport/__init__.py
@@ -5,9 +5,18 @@ raven.transport
 :copyright: (c) 2010-2012 by the Sentry Team, see AUTHORS for more details.
 :license: BSD, see LICENSE for more details.
 """
+# TODO: deprecate this namespace and force non-default (sync + threaded) to
+# manually import/register transports somehow
 from __future__ import absolute_import
 
-from raven.transport.base import (Transport, AsyncTransport, HTTPTransport, GeventedHTTPTransport, TwistedHTTPTransport,  # NOQA
-  TornadoHTTPTransport, RequestsHTTPTransport, UDPTransport, EventletHTTPTransport)  # NOQA
-from raven.transport.exceptions import InvalidScheme, DuplicateScheme  # NOQA
-from raven.transport.registry import TransportRegistry, default_transports  # NOQA
+from raven.transport.base import *  # NOQA
+from raven.transport.eventlet import *  # NOQA
+from raven.transport.exceptions import *  # NOQA
+from raven.transport.gevent import *  # NOQA
+from raven.transport.http import *  # NOQA
+from raven.transport.requests import *  # NOQA
+from raven.transport.registry import *  # NOQA
+from raven.transport.twisted import *  # NOQA
+from raven.transport.threaded import *  # NOQA
+from raven.transport.tornado import *  # NOQA
+from raven.transport.udp import *  # NOQA
