+++ raven/transport/registry.py
@@ -7,12 +7,16 @@ raven.transport.registry
 """
 from __future__ import absolute_import
 
-from raven.transport.base import (
-    HTTPTransport, GeventedHTTPTransport, TwistedHTTPTransport,
-    TornadoHTTPTransport, UDPTransport, EventletHTTPTransport,
-    RequestsHTTPTransport)
+# TODO(dcramer): we really should need to import all of these by default
+from raven.transport.eventlet import EventletHTTPTransport
 from raven.transport.exceptions import DuplicateScheme
+from raven.transport.http import HTTPTransport
+from raven.transport.gevent import GeventedHTTPTransport
+from raven.transport.requests import RequestsHTTPTransport
 from raven.transport.threaded import ThreadedHTTPTransport
+from raven.transport.twisted import TwistedHTTPTransport
+from raven.transport.tornado import TornadoHTTPTransport
+from raven.transport.udp import UDPTransport
 from raven.utils import urlparse
